Api/requirements.txt
pymongo
flask
bcrypt
PyJWT
python-dotenv